Move all artifact storage related classes to one package
Remove unused classes and reformat code
