--- src/main/java/jetbrains/buildServer/investigationsAutoAssigner/utils/ArtifactContent.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2000-2018 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package jetbrains.buildServer.investigationsAutoAssigner.utils;
-
-import java.util.List;
-
-class ArtifactContent {
-  String serverUUID;
-  List<ResponsibilityPersistentInfo> suggestions;
-
-  ArtifactContent(String serverUUID, List<ResponsibilityPersistentInfo> suggestions) {
-    this.serverUUID = serverUUID;
-    this.suggestions = suggestions;
-  }
-}
